Move SimpleCounter to layman.util
